firestore read

+++ lib/data/database.dart
@@ -8,29 +8,50 @@ class DatabaseService {
   final user = FirebaseAuth.instance.currentUser;
 
   void addFavorite(RandomCocktail cocktail) async {
-    final docRef = FirebaseFirestore.instance
-        .collection("users")
-        .doc(user!.uid)
-        .collection("myFavorites")
-        .doc(cocktail.drinkName);
+    if (user!.uid ==
+        FirebaseFirestore.instance.collection("users").doc(user!.uid).id) {
+      final docRef = FirebaseFirestore.instance
+          .collection("users")
+          .doc(user!.uid)
+          .collection("myFavorites")
+          .doc(cocktail.drinkName);
 
-    final json = cocktail.toJson();
-    await docRef.set(json);
+      final json = cocktail.toJson();
+      await docRef.set(json);
+    }
   }
 
   void removeFromFavorite(RandomCocktail cocktail) async {
+    if (user!.uid ==
+        FirebaseFirestore.instance.collection("users").doc(user!.uid).id) {
+      final docRef = FirebaseFirestore.instance
+          .collection("users")
+          .doc(user!.uid)
+          .collection("myFavorites")
+          .doc("Adam Sunrise")
+          .delete()
+          .then(
+            (doc) => print("Document deleted"),
+            onError: (e) => print("Error updating document $e"),
+          );
+      await docRef;
+    } else {
+      null;
+    }
+  }
+
+  Stream<List<RandomCocktail>> myCocktailList(String? userID) {
     final docRef = FirebaseFirestore.instance
         .collection("users")
-        .doc(user!.uid)
+        .doc(userID)
         .collection("myFavorites")
-        .doc("Adam Sunrise")
-        .delete()
-        .then(
-          (doc) => print("Document deleted"),
-          onError: (e) => print("Error updating document $e"),
-        );
-    await docRef;
+        .snapshots()
+        .map((doc) =>
+            doc.docs.map((e) => RandomCocktail.fromJson(e.data())).toList());
+    return docRef;
   }
 }
 
 final db = FutureProvider((ref) => DatabaseService());
+final cocktailListProvider = StreamProvider.family.autoDispose(
+    ((ref, String? userID) => DatabaseService().myCocktailList(userID)));
